v1.0 - First commit :
The 'Baltapp' has three relevant entrypoints (upload pdf, get metadata an get text). The 'Balthclient' is provided to use the API easily.
1) Client's args are correctly checked before to be used. 
2) Then extraction works good. Metadata extracted thanks to PDFreader. And text is extracted thanks to PDFminer. 
3) Data is correctly stored in a  SQLite DB and managed by SqlAlchemy. 

NB : for the moment, the Balthclient is in testing mode. So, execute 'client.py' to test quicly the API with the default provided 'test-text.pdf'. 
Moreover : Code is pretty clean yet (thanks to flake8 and Isort). 

TODO : 
- Type's annotations ;
- Manage the args of the client in the CLI ;
- Add asynchronous tasks (Celery+Redis);
- Add tests (to run Pytest-cov) ;
- Add requirements.txt (freeze);
- Add contract (available with Swagger) ;
- Add the documentation (readme.md).